Beta2.0
Within this version, I added more to the home screen with buttons that take the user to each page. The Home page: welcomes the user to the InnerEase app with the motto "reaching our inner peace" and an uplifting quote or words. There's also a mental check in question; right now, there's nothing, just a question as a placeholder. Eventually, there's going to be a meter or a scale for mood for users to choose. Then there are buttons for the different pages of the app. 
Calendar: just a placeholder right now, there's text that says calendar progress, and letting the user know that they have to complete the check-ins. 
Journal: This page is meant for the users to journal their thoughts. I added the input text box for the user to type in their journal. Inserts the journal into the entry array and saves it. Then the previous journals are shown at the bottom with the timestamp.
Social: Where users can post their thoughts and message other users.
Tools: Ideas for the user to do to help them calm down
Settings: Manages the user's account (notifications and account deletion.