Add files via upload
I updated the HomeScreen to include the mental health check-ins and saved that data within this version.  At first, I was trying to make this a database, but the way I implemented the database was messing with the other aspects of the app. So I decided on a file save for the mood check-ins. I also updated the calendar, social, home, and tools screens. 
HomeScreen 3.0:
 In this version, I added an array of quotes that will randomly display on the screen. Along with that, the option for a user to change the random quote
Goals for next version: My goals for 4.0 are adding a progress widget and goals for the user to set for themselves. Also, to clean up the overall look of the home screen
Journal Screen 3.0: 
 In this version, the Calendar page does not show any activity unless the user enters their mood check in. Once they've filled out their check in it will be displayed on the "calendar" of Check-ins.
Goals:
Display an actual calendar with a mood graph (weekly and monthly, and yearly)  
Journal Goals: The ability to view old journals and to edit and make comments on them 
Social 3.0: turned the placeholder posts into an array that will be displayed in a neat manner
Goals: my goals for this page are to add DMs and the ability for the user to post
Tools 3.0: Within this version, I changed the texts to labels to give it a cleaner look with icons
Goals: I do want to turn these into buttons where they go to a secondary page where the user can get more info on each tab