Add files via upload
This is version 4.0 
What's new?
Rearanged the bottom tabs to have home in the center and maintain sync with the buttons on the home screen. The shared data is JournalStorage and moodStorage, storing the journal entries of the user and the mood that the user is feeling. This was done in a state object to ensure that the data is always fresh. Both use the JSON file system with time stamps. Using the JSON system for the Journal, users can now see all of the previous journals written with the tags of their mood attached as well. I added scroll support on this screen as well as search (its based on what was typed in the journal). For the last part of 4.0 I've added dark mode for the users to experience. 
Future goals:
- Add a progress bar to the home screen 
- add journals to calendar 
-update the social tab